feat: change stars column type to tinyInteger
This commit modifies the `stars` column in the `menu_item_reviews` table. The column type has been changed from `unsignedBigInteger` to `tinyInteger`. This change is made to optimize storage and reflect the expected range of values for a star rating system.

+++ database/migrations/2025_03_09_101718_create_menu_item_reviews_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(MenuItem::class)->constrained()->cascadeOnDelete();
             $table->foreignIdFor(User::class)->constrained()->cascadeOnDelete();
-            $table->unsignedBigInteger('stars')->nullable();
+            $table->tinyInteger('stars')->nullable();
             $table->text('review_text')->nullable();
             $table->timestamps();
         });
